modifide

+++ Besic_programs/Area_Of_cube.py
@@ -1,5 +1,5 @@
 # Python Program to find Volume and Surface Area of a Cube
 
-length = int(4)
+length = float(4)
 area = 6 * (length * length)
 print(" Surface Area of Cube = %.2f"% area)
